Rework CDU
The previous CDU implementation was largely incomplete and unusable.
Despite the incomplete work, this commit introduces several new
features, improvements and fixes, making the CDU closer to the real
aircraft.

Moreover, and as in the real aircraft, the CDUs have been made
independent from each other.

The following pages have been implemented and are reasonably complete:

- IDENT
- POS INIT/POS REF
- DEP/ARR

The following pages are implemented, but are missing features compared
to the real aircraft:

- RTE
- LEGS
- N1 LIMIT
- INIT REF

The following pages have been added, but are still missing many
features:

- FIX
- TAKEOFF
- PROG